+++ src/Statement/WhereStatement.php
@@ -48,14 +48,19 @@ class WhereStatement extends Statement
      *      addCondition('username', 'john')
      * 2. Comparison with custom operator:
      *      addCondition('balance', 100, '>')
-     * 3. SQL fragment:
+     * 3. IN statement:
+     *      addCondition('group', ['admin', 'owner'])
+     * 4. SQL fragment:
      *      addCondition('name LIKE "%john%"')
-     * 4. List of conditions to add:
-     *      addCondition([['balance', 100, '>'], ['user_id', 5]])
-     * 5. Map of equality comparisons:
-     *      addCondition(['username' => 'john', 'user_id' => 5])
-     * 6. List of SQL fragments:
-     *      addCondition(['first_name LIKE "%john%"', 'last_name LIKE "%doe%"'])
+     * 5. List of conditions to add:
+     *      addCondition([['balance', 100, '>'],
+     *                    ['user_id', 5]])
+     * 6. Map of equality comparisons:
+     *      addCondition(['username' => 'john',
+     *                    'user_id' => 5])
+     * 7. List of SQL fragments:
+     *      addCondition(['first_name LIKE "%john%"',
+     *                    'last_name LIKE "%doe%"'])
      *
      * @param array|string $field
      * @param string|bool  $value    condition value (optional)
@@ -67,20 +72,28 @@ class WhereStatement extends Statement
     {
         if (is_array($field) && !$value) {
             foreach ($field as $key => $value) {
-                if (is_array($value)) { // handles #4
+                // handles #5
+                if (is_array($value)) {
                     call_user_func_array([$this, 'addCondition'], $value);
-                } elseif (!is_numeric($key)) { // handles #5
+                // handles #6
+                } elseif (!is_numeric($key)) {
                     $this->addCondition($key, $value);
-                } else { // handles #6
+                // handles #7
+                } else {
                     $this->addCondition($value);
                 }
             }
         } else {
-            // handles #3
+            // handles #4
             $condition = [$field];
 
-            // handles #1 and #2
             if (func_num_args() >= 2) {
+                // handles #3
+                if (is_array($value)) {
+                    $operator = 'IN';
+                }
+
+                // handles #1 and #2
                 $condition[] = $operator;
                 $condition[] = $value;
             }
@@ -101,14 +114,51 @@ class WhereStatement extends Statement
         return $this->conditions;
     }
 
-    protected function buildClause(array $clause)
+    /**
+     * Builds a parameterized and escaped SQL fragment
+     * for a condition that uses our own internal
+     * representation.
+     *
+     * A condition is represented by an array, and can be
+     * have one of the following forms:
+     * i) ['SQL fragment']
+     * ii) ['identifier', '=', 'value']
+     *
+     * @param array $cond
+     *
+     * @return string generated SQL fragment
+     */
+    protected function buildClause(array $cond)
     {
-        // handle NULL values
-        if (count($clause) === 3 && $clause[1] === '=' && $clause[2] === null) {
-            return $this->escapeIdentifier($clause[0]).' IS NULL';
+        // handle SQL fragments
+        if (count($cond) == 1) {
+            return $cond[0];
         }
 
-        return parent::buildClause($clause);
+        // escape the identifier
+        $cond[0] = $this->escapeIdentifier($cond[0]);
+        if (empty($cond[0])) {
+            return '';
+        }
+
+        // handle NULL values
+        if ($cond[1] === '=' && $cond[2] === null) {
+            return $cond[0].' IS NULL';
+        }
+
+        // handle array values, i.e. for IN conditions
+        if (is_array($cond[2])) {
+            foreach ($cond[2] as &$value) {
+                $value = $this->parameterize($value);
+            }
+            $cond[2] = '('.implode(',', $cond[2]).')';
+
+        // otherwise parameterize the value
+        } else {
+            $cond[2] = $this->parameterize($cond[2]);
+        }
+
+        return implode('', $cond);
     }
 
     /**
@@ -118,11 +168,10 @@ class WhereStatement extends Statement
      */
     public function build()
     {
-        $sql = (!$this->having) ? 'WHERE ' : 'HAVING ';
-
+        // build clause from conditions
         $clauses = [];
-        foreach ($this->conditions as $clause) {
-            $clauses[] = $this->buildClause($clause);
+        foreach ($this->conditions as $condition) {
+            $clauses[] = $this->buildClause($condition);
         }
 
         // remove empty values
@@ -131,6 +180,8 @@ class WhereStatement extends Statement
         if (count($clauses) == 0) {
             return '';
         }
+
+        $sql = (!$this->having) ? 'WHERE ' : 'HAVING ';
 
         return $sql.implode(' AND ', $clauses);
     }